ramfs work

+++ enclave/eraise.h
@@ -5,27 +5,27 @@
 #include <errno.h>
 #include <stdio.h>
 
-#define ERAISE(ERRNUM)                                     \
-    do                                                     \
-    {                                                      \
-        ret = ERRNUM;                                      \
-        oel_eraise(__FILE__, __LINE__, __FUNCTION__, ret); \
-        fflush(stdout);                                    \
-        goto done;                                         \
-    }                                                      \
+#define ERAISE(ERRNUM)                                          \
+    do                                                          \
+    {                                                           \
+        ret = ERRNUM;                                           \
+        oel_eraise(__FILE__, __LINE__, __FUNCTION__, (int)ret); \
+        fflush(stdout);                                         \
+        goto done;                                              \
+    }                                                           \
     while (0)
 
-#define ECHECK(ERRNUM)                                         \
-    do                                                         \
-    {                                                          \
-        int _r_ = ERRNUM;                                      \
-        if (_r_ != 0)                                          \
-        {                                                      \
-            ret = _r_;                                         \
-            oel_eraise(__FILE__, __LINE__, __FUNCTION__, ret); \
-            goto done;                                         \
-        }                                                      \
-    }                                                          \
+#define ECHECK(ERRNUM)                                              \
+    do                                                              \
+    {                                                               \
+        int _r_ = ERRNUM;                                           \
+        if (_r_ != 0)                                               \
+        {                                                           \
+            ret = _r_;                                              \
+            oel_eraise(__FILE__, __LINE__, __FUNCTION__, (int)ret); \
+            goto done;                                              \
+        }                                                           \
+    }                                                               \
     while (0)
 
 void oel_eraise(
